+++ tests/test_execution_engine.py
@@ -3,7 +3,7 @@ import os
 import shutil
 import json
 from quanta_tissu.tisslm.core.execution_engine import ExecutionEngine, ToolRegistry, TissSecurityError, TissAssertionError
-from quanta_tissu.tisslm.integrations.tisslang_parser import TissLangParser
+from quanta_tissu.tisslm.parser.tisslang_parser import TissLangParser
 
 class TestExecutionEngine(unittest.TestCase):
